refactoring JSON API POST and reorganizing the fetch module in the worker
